[CodeClimate fix] [Backend] moved create-pia-mock to separate file
